requirements.txt
Django
django-allauth
python-dotenv
django-crispy-forms